+++ data/Duke.txt
@@ -2,5 +2,3 @@ T | 1 | read book
 E | 1 | project meeting | Aug 6th 2-4pm
 T | 1 | join sports club
 T | 0 | jump
-D | 0 | return book | 2019-12-02T18:00
-T | 1 | sleep
