+++ src/main/java/byu/FileReader.java
@@ -1,6 +1,9 @@
+package byu;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.util.Scanner;
+import task.*;
 
 public class FileReader {
