+++ src/main/java/byu/Instruction.java
@@ -1,3 +1,5 @@
+package byu;
+
 public enum Instruction {
     LIST, MARK, UNMARK, TODO, DEADLINE, EVENT, DELETE, BYE
 }
