+++ src/main/java/commands/ByeCommand.java
@@ -1,3 +1,7 @@
+package commands;
+import byu.*;
+import exceptions.InvalidIndex;
+
 public class ByeCommand extends Command {
 
     public void execute(ToDoList list, Ui ui) throws InvalidIndex {
