+++ src/main/java/commands/Command.java
@@ -1,3 +1,7 @@
+package commands;
+import byu.*;
+import exceptions.InvalidIndex;
+
 public abstract class Command {
 
     public abstract void execute(ToDoList list, Ui ui) throws InvalidIndex;
