+++ src/main/java/commands/NextCommand.java
@@ -1,3 +1,6 @@
+package commands;
+import byu.*;
+
 public class NextCommand extends Command {
 
     public void execute(ToDoList list, Ui ui) {
