+++ src/main/java/exceptions/InvalidInstruction.java
@@ -1,3 +1,5 @@
+package exceptions;
+
 public class InvalidInstruction extends DukeException {
 
     public InvalidInstruction() {
